[v0.91.2][maintainability] Split adl/src/runtime_v2/observatory_flagship.rs by responsibility (Closes #3156)
